requirement.txt
pydantic>=1.8.2
pyyaml>=5.3.1
requests>=2.22.0
rich>=10.16.0
